move TN guard to better location; make guard print error message rather than throwing error
